% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_atree.R
\name{active_nodes_plot}
\alias{active_nodes_plot}
\title{Create the active nodes in memory plot}
\usage{
active_nodes_plot(Application = NULL, Atree = NULL, step = 100)
}
\arguments{
\item{Application}{starvz_data Application trace data}

\item{Atree}{starvz_data Atree elimination tree trace data}
}
\value{
A ggplot object
}
\description{
Use starvz_data to create a line plot of the number of active nodes per type
along the application execution time
}
\examples{
active_nodes_plot(data$Application, data$Atree, step=100)
}
