% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_atree.R
\name{atree_temporal_plot}
\alias{atree_temporal_plot}
\title{Create the elimination tree plot with some options in the config file}
\usage{
atree_temporal_plot(Application = NULL, Atree = NULL, step = 100)
}
\arguments{
\item{Application}{starvz_data Application trace data}

\item{Atree}{starvz_data Atree elimination tree trace data}

\item{step}{size in milliseconds for the time aggregation step}
}
\value{
A ggplot object
}
\description{
Use starvz_data to create a representation of the elimination tree structure
considering the initialization, communication and computational tasks. These
representations can be controlled in the configuration file.
}
\examples{
atree_temporal_plot(data$Application, data$Atree, step=100)
}
