% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_khills.R
\name{geom_khills}
\alias{geom_khills}
\title{Khills tracking rectangle layer}
\usage{
geom_khills(factor = 1, slice_size = 100, levels = NULL, ...)
}
\arguments{
\item{factor}{Numeric factor value to step iterations}

\item{slice_size}{Numeric aggregation step size for time slicing}

\item{levels}{Character vector specifying the order of tasks}

\item{...}{Additional arguments passed to geom_rect}
}
\value{
A ggplot2 layer
}
\description{
Khills tracking rectangle layer
}
