% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_khills.R
\name{geom_khills_makespan}
\alias{geom_khills_makespan}
\title{Khills makespan label layer}
\usage{
geom_khills_makespan(size = 5)
}
\arguments{
\item{size}{Numeric font size for the label}
}
\value{
A ggplot2 layer
}
\description{
Khills makespan label layer
}
