% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_memory.R
\name{handles_presence_states}
\alias{handles_presence_states}
\title{Computes presence of handles over resources}
\usage{
handles_presence_states(data)
}
\arguments{
\item{data}{starvz_data with trace data}
}
\value{
Time-Step aggregated handle presences
}
\description{
Use for precomputation of other memory-related functions
}
\examples{
\donttest{
handles_presence_states(starvz_sample_lu)
}
}
