% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_memory.R
\name{multiple_snaps}
\alias{multiple_snaps}
\title{Create multiple snapshot of memory}
\usage{
multiple_snaps(
  data = NULL,
  start = 0,
  end = 1000,
  step = 100,
  path = ".",
  scale = 8,
  width = 4,
  height = 3
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{start}{start time}

\item{end}{end time}

\item{step}{between snaps}

\item{path}{path to save files}

\item{scale}{for ggsave}

\item{width}{for ggsave}

\item{height}{for ggsave}
}
\value{
A ggplot object
}
\description{
Create multiple visualizations of memory
Useful for continuing views
}
\examples{
\dontrun{
multiple_snaps(data = starvz_sample_lu, 100, 200, 10, ".")
}
}
