% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_atree.R
\name{nodes_memory_usage_plot}
\alias{nodes_memory_usage_plot}
\title{Create the node memory usage plot}
\usage{
nodes_memory_usage_plot(Application = NULL, step = 100)
}
\arguments{
\item{Application}{starvz_data Application trace data}

\item{step}{size in milliseconds for the time aggregation step}
}
\value{
A ggplot object
}
\description{
Use starvz_data to create a line plot of the memory usage in MB of
active nodes along the application execution time
}
\examples{
active_nodes_plot(data$Application, step=100)
}
