% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_metrics.R
\name{panel_abe_solution}
\alias{panel_abe_solution}
\title{Create a plot with the solution computed by ABE}
\usage{
panel_abe_solution(data)
}
\arguments{
\item{data}{starvz_data with trace data
#' @return A ggplot object}
}
\description{
Plot per-node and per-tasktype repartion among resource types
}
\examples{
\donttest{
panel_abe_solution(data = starvz_sample_lu)
}
}
