% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_atree.R
\name{panel_atree}
\alias{panel_atree}
\title{Create the elimination tree plot with some options in the config file}
\usage{
panel_atree(
  data = NULL,
  step = 100,
  legend = TRUE,
  zoom = FALSE,
  computation = TRUE,
  pruned = TRUE,
  initialization = TRUE,
  x_start = data$config$limits$start,
  x_end = data$config$limits$end,
  communication = TRUE,
  anomalies = TRUE
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{step}{size in milliseconds for the time aggregation step}

\item{legend}{enable/disable panel legend}

\item{zoom}{enable/disable vertical zoom in the tree structure}

\item{computation}{enable/disable computations representations in the tree}

\item{pruned}{enable/disable pruned computations representations in the tree}

\item{initialization}{enable/disable initialization tasks representation}

\item{x_start}{X-axis start value}

\item{x_end}{X-axis end value}

\item{communication}{enable/disable communication tasks representation}

\item{anomalies}{enable/disable anomalies tasks representation}
}
\value{
A ggplot object
}
\description{
Use starvz_data to create a representation of the elimination tree structure
considering the initialization, communication and computational tasks. These
representations can be controlled in the configuration file.
}
\examples{
#panel_atree(data, step=10)
#panel_atree(data, step=20, communication=FALSE, initialization=FALSE)
}
