% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_atree.R
\name{panel_atree_structure}
\alias{panel_atree_structure}
\title{Create the elimination tree structure plot along time}
\usage{
panel_atree_structure(data = NULL, end_arrow = "ParentEnd")
}
\arguments{
\item{data}{starvz_data with trace data}

\item{end_arrow}{behavior of the end arrow [ParentEnd, ComputationEnd]}
}
\value{
A ggplot object
}
\description{
Use Atree and Application data to create the elimination tree
structure plot in a ggplot object and return it
}
\examples{
\dontrun{
panel_atree_structure(starvz_sample_lu)
}
}
