% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_khills.R
\name{panel_khills}
\alias{panel_khills}
\title{Create a standard khills panel for application traces}
\usage{
panel_khills(
  data = NULL,
  factor = data$config$khills$factor,
  slice_size = data$config$khills$slice_size,
  makespan_size = data$config$khills$makespan_size,
  levels = data$config$khills$levels
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{factor}{factor value to step iterations}

\item{slice_size}{aggregation step size for time slicing}

\item{makespan_size}{font size for the total execution label}

\item{levels}{Character vector specifying the order of tasks}
}
\value{
A ggplot object
}
\description{
Create a standard khills panel for application traces
}
