% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_kchart.R
\name{panel_kiteration}
\alias{panel_kiteration}
\title{Create a special chart for applications with iterations}
\usage{
panel_kiteration(
  data = NULL,
  middle_lines = data$config$kiteration$middlelines,
  base_size = data$config$base_size,
  expand_x = data$config$expand,
  legend = data$config$kiteration$legend,
  x_start = data$config$limits$start,
  x_end = data$config$limits$end,
  per_node = data$config$kiteration$pernode,
  sub_ite = data$config$kiteration$subite
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{middle_lines}{plot a middle line}

\item{base_size}{base_size base font size}

\item{expand_x}{expand size for scale_x_continuous padding}

\item{legend}{enable/disable legends}

\item{x_start}{X-axis start value}

\item{x_end}{X-axis end value}

\item{per_node}{Create node facets}

\item{sub_ite}{Use Subiteration as Y}
}
\value{
A ggplot object
}
\description{
Plot iterations Y over Time X
}
\examples{
panel_kiteration(data = starvz_sample_lu)
}
