% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_lackready.R
\name{panel_lackready}
\alias{panel_lackready}
\title{Shows if the runtimes is lacking ready tasks}
\usage{
panel_lackready(
  data = NULL,
  x_start = data$config$limits$start,
  x_end = data$config$limits$end
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{x_start}{X-axis start value}

\item{x_end}{X-axis end value}
}
\value{
A ggplot object
}
\description{
Plot a bar over time that shows when the runtime is lacking ready tasks
}
\examples{
panel_lackready(data = starvz_sample_lu)
}
