% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_memory.R
\name{panel_memory_heatmap}
\alias{panel_memory_heatmap}
\title{Heatmap of memory presence}
\usage{
panel_memory_heatmap(
  data,
  legend = data$config$memory_heatmap$legend,
  base_size = data$config$base_size,
  expand_x = data$config$expand
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{legend}{enable/disable legends}

\item{base_size}{base_size base font size}

\item{expand_x}{expand size for scale_x_continuous padding}
}
\value{
A ggplot object
}
\description{
Visualizate the presence of memory handles across memory managers
}
\examples{
\donttest{
panel_memory_heatmap(data = starvz_sample_lu)
}
}
