% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_memory.R
\name{panel_memory_snap}
\alias{panel_memory_snap}
\title{Create a snapshot of memory}
\usage{
panel_memory_snap(data, selected_time, step, tasks_size = 30)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{selected_time}{time}

\item{step}{for discreate events}

\item{tasks_size}{size of tasks in the visualization}
}
\value{
A ggplot object
}
\description{
Visualizate memory in a specific time
To accelerate the process:
data$handle_states <- handles_presence_states(data)
data$handle_gantt_data <- pre_handle_gantt(data)
To Select time:
handles_gantt(data, JobId=c(jobid))
snap_data <- pre_snap(data, data$handle_states)
memory_snap(snap_data, 1000, tasks_size=200, step=1)
}
\examples{
panel_memory_snap(data = starvz_sample_lu, 100, 10)
}
