% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_util.R
\name{panel_model_gflops}
\alias{panel_model_gflops}
\title{Create the diagnostig plot for the regression model}
\usage{
panel_model_gflops(data,
         freeScales = TRUE, model_type = "LOG_LOG")
}
\arguments{
\item{data}{starvz_data with trace data}

\item{freeScales}{free X,Y scales for each task and resource type combination}

\item{model_type}{Choose the regression model type to use}
}
\value{
A ggplot object
}
\description{
Use the starvz Application data to observe how the regression model used
in the task anomaly classification fits the data.
}
\examples{
\dontrun{
panel_model_gflops(data = starvz_sample_sample)
}
}
