% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_mpi.R
\name{panel_mpistate}
\alias{panel_mpistate}
\title{Create a space-time view of MPI controlers}
\usage{
panel_mpistate(
  data = NULL,
  legend = data$config$mpibandwidth$legend,
  base_size = data$config$base_size,
  expand_x = data$config$expand,
  x_start = data$config$limits$start,
  x_end = data$config$limits$end,
  y_start = 0,
  y_end = data$config$mpibandwidth$limit
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{legend}{enable/disable legends}

\item{base_size}{base_size base font size}

\item{expand_x}{expand size for scale_x_continuous padding}

\item{x_start}{X-axis start value}

\item{x_end}{X-axis end value}

\item{y_start}{Y-axis start value}

\item{y_end}{Y-axis end value}
}
\value{
A ggplot object
}
\description{
Create a space-time view of MPI controlers
}
\examples{
panel_mpistate(data=starvz_sample_lu)
}
