% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_node_summary.R
\name{panel_node_summary}
\alias{panel_node_summary}
\title{Create a bar plot with node information}
\usage{
panel_node_summary(
  data,
  legend = data$config$summary_nodes$legend,
  base_size = data$config$base_size,
  expand_x = data$config$expand,
  x_start = data$config$limits$start,
  x_end = data$config$limits$end
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{legend}{enable/disable legends}

\item{base_size}{base_size base font size}

\item{expand_x}{expand size for scale_x_continuous padding}

\item{x_start}{X-axis start value}

\item{x_end}{X-axis end value}
}
\value{
A ggplot object
}
\description{
Bar plot with makespan and abe per node
}
\examples{
panel_node_summary(data = starvz_sample_lu)
}
