% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_atree.R
\name{panel_nodememuse}
\alias{panel_nodememuse}
\title{Create the node memory usage plot}
\usage{
panel_nodememuse(data = NULL, step = 100, aggregation = FALSE, legend = TRUE)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{step}{size in milliseconds for the time aggregation step}

\item{aggregation}{enable/disable time aggregation for the plot}

\item{legend}{enable/disable plot legends}
}
\description{
Use starvz_data to create a line plot of the memory usage in MB of
active nodes along the application execution time
}
\examples{
panel_nodememuse(data$Application, step=100)
}
