% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_pmtool.R
\name{panel_pmtool_kiteration}
\alias{panel_pmtool_kiteration}
\title{Create a special chart for applications with iterations with PMtool data}
\usage{
panel_pmtool_kiteration(
  data = NULL,
  legend = data$config$pmtool$kiteration$legend,
  x_start = data$config$limits$start,
  x_end = data$config$limits$end
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{legend}{enable/disable legends}

\item{x_start}{X-axis start value}

\item{x_end}{X-axis end value}
}
\value{
A ggplot object
}
\description{
Plot iteraionts Y over Time X of PMtool data
}
\examples{
panel_pmtool_kiteration(data = starvz_sample_lu)
}
