% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_pmtool.R
\name{panel_pmtool_st}
\alias{panel_pmtool_st}
\title{Create a space time visualization of pmtool application as a Gantt chart}
\usage{
panel_pmtool_st(
  data = NULL,
  legend = data$config$pmtool$state$legend,
  base_size = data$config$base_size,
  expand_x = data$config$expand,
  x_start = data$config$limits$start,
  x_end = data$config$limits$end
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{legend}{enable/disable legends}

\item{base_size}{base_size base font size}

\item{expand_x}{expand size for scale_x_continuous padding}

\item{x_start}{X-axis start value}

\item{x_end}{X-axis end value}
}
\value{
A ggplot object
}
\description{
Use the PMTOOL Application trace data to plot the task computations by ResourceId
over the execution time.
}
\examples{
panel_pmtool_st(data = starvz_sample_lu)
}
