% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_util.R
\name{panel_resource_usage_task}
\alias{panel_resource_usage_task}
\title{Plot resource utilization using tasks as color}
\usage{
panel_resource_usage_task(
  data = NULL,
  step = NULL,
  legend = FALSE,
  x_start = data$config$limits$start,
  x_end = data$config$limits$end
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{step}{size in milliseconds for the time aggregation step}

\item{legend}{enable/disable plot legends}

\item{x_start}{X-axis start value}

\item{x_end}{X-axis end value}
}
\value{
A ggplot object
}
\description{
Use data Application to create a panel of the total resource utilization
that helps to observe the time related resource utilization by task
}
\examples{
\donttest{
panel_resource_usage_task(data = starvz_sample_lu)
}
}
