% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_states_chart.R
\name{panel_st}
\alias{panel_st}
\title{Create a space time visualization of the application as a Gantt chart}
\usage{
panel_st(
  data,
  agg = data$config$st$aggregation$active,
  agg_met = data$config$st$aggregation$method
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{agg}{boolean Active or not aggregation}

\item{agg_met}{Aggregation method, possible: static, dynamic, nodes}
}
\value{
A ggplot object
}
\description{
Use the Application trace data to plot the task computations by ResourceId
over the execution time. It will select the aggregation mode if requested.
}
\examples{
panel_st(data = starvz_sample_lu)
}
