% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_aggregation.R
\name{panel_st_agg_dynamic}
\alias{panel_st_agg_dynamic}
\title{Create a space time visualization as a Gantt chart with dynamic aggregation}
\usage{
panel_st_agg_dynamic(
  data = NULL,
  x_start = data$config$limits$start,
  x_end = data$config$limits$end
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{x_start}{X-axis start value}

\item{x_end}{X-axis end value}
}
\value{
A ggplot object
}
\description{
Use the any state trace data to plot the task computations by ResourceId
over the execution time. Dynamic aggregating by selecting the best time-step
}
\examples{
#panel_st_agg_dynamic(data = starvz_sample_lu)
}
