% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_states_chart.R
\name{panel_st_runtime}
\alias{panel_st_runtime}
\title{Create a space time visualization of the runtime as a Gantt chart}
\usage{
panel_st_runtime(data,
             agg = data$config$starpu$aggregation$active)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{agg}{Active or not static aggregation}
}
\value{
A ggplot object
}
\description{
Use the runtime trace data to plot the task computations by ResourceId
over the execution time. It will select the aggregation mode if requested,
only static aggregation is available for runtime.
}
\examples{
panel_st_runtime(data = starvz_sample_lu)
}
