% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_util.R
\name{panel_title}
\alias{panel_title}
\title{Create the title of StarVZ plot}
\usage{
panel_title(data)
}
\arguments{
\item{data}{starvz_data with trace data}
}
\value{
A ggplot object
}
\description{
Use the directory of traces name to create a plot title
}
\examples{
panel_title(data = starvz_sample_lu)
}
