% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_atree.R
\name{panel_utiltreenode}
\alias{panel_utiltreenode}
\title{Create the resource utilization by tree node plot}
\usage{
panel_utiltreenode(data = NULL, step = 100)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{step}{size in milliseconds for the time aggregation step}
}
\value{
A ggplot object
}
\description{
Use starvz_data Application and Atree to create a plot that shows the
total resource utilization, painted by tree node using geom_ribbon. The
colors are reused between nodes, not tied to as specific tree node.
}
\examples{
panel_utiltreenode(data=starvz_data, step=100)
}
