% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starvz_data.R
\name{plot.starvz_data}
\alias{plot.starvz_data}
\title{Plot starvz_data}
\usage{
\method{plot}{starvz_data}(x, ...)
}
\arguments{
\item{x}{A starvz_data}

\item{...}{optional}
}
\value{
Nothing
}
\description{
Plot starvz_data
}
\examples{
\donttest{
plot(starvz_sample_lu)
}
}
