% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_memory.R
\name{pre_handle_gantt}
\alias{pre_handle_gantt}
\title{Pre-Computation for memory handles panel}
\usage{
pre_handle_gantt(data, name_func = NULL)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{name_func}{function to give names to handles}
}
\value{
Pre-Computated data for panel_handles
}
\description{
Use for precomputation of memory handles panel
}
\examples{
\donttest{
pre_handle_gantt(data = starvz_sample_lu)
}
}
