% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starvz_data.R
\name{print.starvz_data}
\alias{print.starvz_data}
\title{Print starvz_data}
\usage{
\method{print}{starvz_data}(x, ...)
}
\arguments{
\item{x}{A starvz_data}

\item{...}{optional}
}
\value{
Nothing
}
\description{
Print starvz_data
}
\examples{
print(starvz_sample_lu)
}
