% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_atree.R
\name{resource_utilization_tree_node}
\alias{resource_utilization_tree_node}
\title{Create the node memory usage plot}
\usage{
resource_utilization_tree_node(
  Application = NULL,
  Atree = NULL,
  step = 100,
  group_pruned = FALSE,
  performance_metric = "Time"
)
}
\arguments{
\item{Application}{starvz application data}

\item{Atree}{starvz elimination tree data}

\item{step}{size in milliseconds for the time aggregation step}

\item{group_pruned}{aggregate computations of the same parent pruned nodes}

\item{performance_metric}{Performance metric to save in Value1 [Time, GFlops]}
}
\description{
Use starvz_data to create a line plot of the memory usage in MB of
active nodes along the application execution time
}
