% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_atree.R
\name{resource_utilization_tree_node_plot}
\alias{resource_utilization_tree_node_plot}
\title{Create the resource utilization by tree node plot}
\usage{
resource_utilization_tree_node_plot(
  Application = NULL,
  Atree = NULL,
  step = 100
)
}
\arguments{
\item{Application}{starvz_data Application trace data}

\item{Atree}{starvz_data Atree elimination tree trace data}

\item{step}{size in milliseconds for the time aggregation step}
}
\value{
A ggplot object
}
\description{
Use starvz_data Application and Atree to create a plot that shows the
total resource utilization, painted by tree node using geom_ribbon. The
colors are reused between nodes, not tied to as specific tree node.
}
\examples{
resource_utilization_tree_node_plot(data$Application, data$Atree, step=100)
}
