% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_khills.R
\name{scale_khills}
\alias{scale_khills}
\title{Standard scales and labels for khills panels}
\usage{
scale_khills(colors = NULL)
}
\arguments{
\item{colors}{A named character vector of colors}
}
\value{
A list of ggplot2 scales and labels
}
\description{
Standard scales and labels for khills panels
}
