% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2.R
\name{starvz_assemble}
\alias{starvz_assemble}
\title{Assemble multiple StarVZ panel lists}
\usage{
starvz_assemble(
  ...,
  config = NULL,
  remove_Y_info = TRUE,
  remove_legends = TRUE
)
}
\arguments{
\item{...}{Panel list or list of panel lists}

\item{config}{StarVZ configurations for determaning panels heights}

\item{remove_Y_info}{remove Y labels for a second and subsequent list of panels}

\item{remove_legends}{remove legends for a second and subsequent list of panels}
}
\value{
The ggplot plot
}
\description{
Take a panel list, or a list of list of panels and assemble it
}
\examples{
\donttest{
starvz_assemble(starvz_plot_list(starvz_sample_lu),
  config = starvz_sample_lu$config
)
}
}
