% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2.R
\name{starvz_guided_plot}
\alias{starvz_guided_plot}
\title{Create a StarVZ plot automaticaly computing height}
\usage{
starvz_guided_plot(data, name)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{name}{Path for saved image}
}
\value{
The ggplot plot
}
\description{
Use data to create a starvz plot and save in name
}
\examples{
#starvz_guided_plot("data/", "file.png")
}
