% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1.R
\name{starvz_phase1_read_write}
\alias{starvz_phase1_read_write}
\title{Execute StarVZ phase 1}
\usage{
starvz_phase1_read_write(
  directory = ".",
  app_states_fun = NULL,
  state_filter = 0,
  whichApplication = NULL,
  input.parquet = "1"
)
}
\arguments{
\item{directory}{Directory of csv files}

\item{app_states_fun}{Function to determine application}

\item{state_filter}{Type of filder}

\item{whichApplication}{Name of Application}

\item{input.parquet}{Use or not of parquet files}
}
\value{
ggplot object with all starvz plots
}
\description{
Convert CSVs to StarVZ Files
}
