% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2.R
\name{starvz_plot}
\alias{starvz_plot}
\title{Make a StarVZ plot}
\usage{
starvz_plot(data = NULL)
}
\arguments{
\item{data}{starvz_data class with $config}
}
\value{
ggplot object with all starvz plots
}
\description{
Create a StarVZ plot considering the data suplied
}
\examples{
# starvz_plot(data)
}
