% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2.R
\name{starvz_plot_list}
\alias{starvz_plot_list}
\title{Generate the StarVZ Plots}
\usage{
starvz_plot_list(data = NULL)
}
\arguments{
\item{data}{starvz_data with trace data}
}
\value{
A list of ggplot plots
}
\description{
Use data to create the list of StarVZ plots
}
\examples{
\donttest{
starvz_plot_list(starvz_sample_lu)
}
}
