% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starvz_data.R
\docType{data}
\name{starvz_sample_lu}
\alias{starvz_sample_lu}
\title{Small StarVZ data of LU Factorization}
\format{
An object of class \code{starvz_data} of length 21.
}
\source{
Chameleon+StarPU
}
\usage{
starvz_sample_lu
}
\description{
A small StarVZ data object obtained from Chameleon+StarPU LU Factorization
}
\keyword{datasets}
