% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{starvz_selective_read}
\alias{starvz_selective_read}
\title{Read selective starvz trace files}
\usage{
starvz_selective_read(directory = ".", config_file = NULL)
}
\arguments{
\item{directory}{Directory path of trace files}

\item{config_file}{Path for configuration yaml file}
}
\value{
The starvz_data with all tables
}
\description{
Read the directory of trace files (feather or parquet) and
the configfuration file, and return a starvz_data class with only
data that will be used in active plots in cofiguration file
}
\examples{
starvz_selective_read("folder_to_parquet_files/")
starvz_selective_read(directory = "folder_to_parquet_files/", config_file = "path_to_config.yaml")
starvz_selective_read() # Read current directory
}
