% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starvz_data.R
\name{starvz_set_log}
\alias{starvz_set_log}
\title{Active internal debug logs}
\usage{
starvz_set_log(state)
}
\arguments{
\item{state}{Active or not logs}
}
\value{
Nothing
}
\description{
Active internal debug logs
}
\examples{
starvz_set_log(FALSE)
}
