% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_atree.R
\name{panel_activenodes}
\alias{panel_activenodes}
\title{Create the active nodes in memory plot}
\usage{
panel_activenodes(
  data = NULL,
  step = data$config$activenodes$aggregation$step,
  aggregation = data$config$activenodes$aggregation$active,
  x_start = data$config$limits$start,
  x_end = data$config$limits$end,
  legend = data$config$activenodes$legend
)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{step}{size in milliseconds for the time aggregation step}

\item{aggregation}{enable/disable time aggregation for the plot}

\item{x_start}{X-axis start value}

\item{x_end}{X-axis end value}

\item{legend}{enable/disable plot legends}
}
\value{
A ggplot object
}
\description{
Use starvz_data to create a line plot of the number of active nodes per type
along the application execution time
}
\examples{
\dontrun{
panel_activenodes(data = starvz_sample_lu, step = 100)
}
}
