% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_memory.R
\name{panel_memory_snap}
\alias{panel_memory_snap}
\title{Create a snapshot of memory}
\usage{
panel_memory_snap(data, selected_time, step, tasks_size = 30)
}
\arguments{
\item{data}{starvz_data with trace data}

\item{selected_time}{time}

\item{step}{for discreate events}

\item{tasks_size}{size of tasks in the visualization}
}
\value{
A ggplot object
}
\description{
Visualizate memory in a specific time
}
\examples{
\donttest{
panel_memory_snap(data = starvz_sample_lu, 100, 10)
}
}
