% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2.R
\name{starvz_plot}
\alias{starvz_plot}
\title{Make a StarVZ plot}
\usage{
starvz_plot(
  data = NULL,
  name = NULL,
  save = FALSE,
  guided = data$config$guided$active,
  dpi = 120
)
}
\arguments{
\item{data}{starvz_data class with $config}

\item{name}{Path for saved image}

\item{save}{call ggplot to save the image}

\item{guided}{compute ideal figure height}

\item{dpi}{dpi for ggsave}
}
\value{
ggplot object with all starvz plots
}
\description{
Create a StarVZ plot considering the data supplied
}
\examples{
\donttest{
starvz_plot(starvz_sample_lu)
}
}
