% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{starvz_read}
\alias{starvz_read}
\title{Read starvz trace files}
\usage{
starvz_read(directory = ".", config_file = NULL, selective = TRUE)
}
\arguments{
\item{directory}{Directory path of trace files}

\item{config_file}{Path for configuration yaml file}

\item{selective}{if True, only read data needed for creating panels activated
in config}
}
\value{
The starvz_data with all tables
}
\description{
Read the directory of trace files (feather or parquet) and
the configuration file, and return a starvz_data class used
in starvz functions
}
\examples{
starvz_read("folder_to_parquet_files/")
starvz_read(directory = "folder_to_parquet_files/", config_file = "path_to_config.yaml")
starvz_read() # Read current directory
}
