% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starvz_data.R
\name{summary.starvz_data}
\alias{summary.starvz_data}
\title{Summary starvz_data}
\usage{
\method{summary}{starvz_data}(object, ...)
}
\arguments{
\item{object}{A starvz_data}

\item{...}{optional}
}
\value{
Nothing
}
\description{
Summary starvz_data
}
\examples{
summary(starvz_sample_lu)
}
